==> 3. Data Analysis.Rmd <==
---
title: "Data Analysis"
author: "Himanshu"
date: "April 6, 2019"
output: html_document
---
```{r}
library(dplyr)
library(caret)
library(lubridate)
library(stringr)
library(tidyr)
library(corrplot)
library(ggcorrplot)
library(corrgram)
library(DescTools)
require(grid)
library(ggplot2)
library(gridExtra)
```

```{r}
# Loading data
df <- read.csv("Data.csv", header = TRUE)

```

```{r}
# Removing Primary Keys:
rm_col <- c("X", "id", "member_id", "profit")
df<- df[, !(colnames(df) %in% rm_col), drop = FALSE]
str(df)
```
```{r}
#Converting date variables into date type

df$earliest_cr_line <- as.Date(df$earliest_cr_line)
df$issue_d <- as.Date(df$issue_d)
```
```{r}
#checking percentage of NA in data
tmp = sort(sapply(df, function(x) sum(length(which(is.na(x)))))/nrow(df),decreasing = TRUE)
tmp           
```

```{r}
#earliest_cr_line has some NA's, removing NA's containing records

df2 <- na.omit(df1, cols = c("earliest_cr_line"))

```

```{r}
# Fetching numeric and categorical variables
numVars<- names(which(sapply(df2, is.numeric)))
catvars <- names(which(sapply(df2, is.factor)))
```


```{r}
# correlation plot of all Numeric variables


Primary_numeric <- c("funded_amnt", "funded_amnt_inv","loan_amnt", "int_rate","installment", "open_acc", "dti" , "annual_inc", "pub_rec", "total_acc")

# Correlation plot for Primary data source variables
corrplot(cor(df2[,Primary_numeric],use="na.or.complete"), type = "lower", method = "circle")


```
```{r}

a <- cor(as.matrix(df2[,Primary_numeric]), use = "pairwise.complete.obs")
a[lower.tri(a,diag=TRUE)]=NA    #Prepare to drop duplicates and meaningless information
a <- as.data.frame(as.table(a)) #Turn into a 3-column table
a <- na.omit(a)                 #Get rid of the junk we flagged above
a <- a[order(-abs(a$Freq)),]    #Sort by highest correlation (whether +ve or -ve)
head(a, n = 20)
# loan_amount, funded amount and funded amount_inv have perfect correlation
```

```{r}
# Correlation plot for Secondary data source variables

Secondary_numeric <- c("Gross.domestic.product..GDP..by.state", "Compensation.of.employees", "Gross.operating.surplus", "Per.capita.real.GDP.by.state" , "Quantity.indexes.for.real.GDP.by.state",
              "Real.GDP.by.state" ,  "Subsidies",  "Taxes.on.production.and.imports.less.subsidies", 
              "Taxes.on.production.and.imports")

corrplot(cor(df2[,Secondary_numeric],use="na.or.complete"), type = "lower", method = "circle")

```


```{r}
a <- cor(as.matrix(df2[,Secondary_numeric]), use = "pairwise.complete.obs")
a[lower.tri(a,diag=TRUE)]=NA    #Prepare to drop duplicates and meaningless information
a <- as.data.frame(as.table(a)) #Turn into a 3-column table
a <- na.omit(a)                 #Get rid of the junk we flagged above
a <- a[order(-abs(a$Freq)),]    #Sort by highest correlation (whether +ve or -ve)
head(a, n = 20)
# Real GDP and GDP by state has high correlaion
```



```{r}

#Univariate Analysis

Desc(df2$loan_amnt, main = "Loan amount distribution", plotit = TRUE)
```
```{r}
Desc(df2$purpose, main = "Loan purposes", plotit = TRUE)

# most of the borrower took loan for debt consolidation
```
```{r}
Desc(df2$term, main = "Term", plotit = TRUE)
```



```{r}
Desc(df2$grade, main = "Loan grades", plotit = TRUE)
Desc(df2$sub_grade, main= "Loan sub grades", plotit= TRUE)
```
```{r}
# Removing unsed factors in home ownership
df2$home_ownership <- factor(df2$home_ownership)
table(df2$home_ownership)
df2<- df2 %>% filter(home_ownership=="MORTGAGE"|home_ownership=="OWN"|home_ownership=="RENT")
df2$home_ownership <- factor(df2$home_ownership)
table(df2$home_ownership)
Desc(df2$home_ownership, main = "Home Ownership", plotit = TRUE)
```
```{r}
Desc(df2$emp_title, main = "Employee Title", plotit = TRUE)

# too many others values, should be removed
```
```{r}
Desc(df2$application_type, main = "Application type", plotit = TRUE)


```
```{r}
Desc(df2$initial_list_status, main= "Initial list status", plotit= TRUE)
```
```{r}
Desc(df2$verification_status, main= "Verification status", plotit= TRUE)
```


```{r}
# Multivariate analysis

# Plotting total number of credit lines vs open credit lines

p1 <- ggplot(df2, aes(x = open_acc)) + geom_histogram(aes(fill = loan_status), position = "dodge", bins = 10) + labs(x= "Open Credit Lines")
p2<- ggplot(df2, aes(x = total_acc)) + geom_histogram(aes(fill = loan_status), position = "dodge", bins = 10) + labs(x= "Total Credit Lines")

# Legend function: Fetching legend from ggplot
g_legend<-function(a.gplot){
  tmp <- ggplot_gtable(ggplot_build(a.gplot))
  leg <- which(sapply(tmp$grobs, function(x) x$name) == "guide-box")
  legend <- tmp$grobs[[leg]]
  return(legend)}

mylegend<-g_legend(p1) # use legend of plot 1

# grid arrange : To plot multiple plots on same graph
grid.arrange(
  p1 +theme(legend.position="none"), # remoing individual plot legend
  p2+theme(legend.position="none"),  # remoing individual plot legend
  mylegend,
  nrow = 1,
  top = "Credit line distribution",
  bottom = textGrob(
    " ",
    gp = gpar(fontface = 3, fontsize = 9),
    hjust = 1,
    x = 1
  )
)

# more number of credit lines are open for non defaulters 
```
```{r}
# Number of derogatory public records - pub_rec

ggplot(df2, aes(x = pub_rec)) + geom_histogram(aes(fill = loan_status), position = "dodge", bins = 10)
```
```{r}
p1 <- ggplot(df2, aes(x = loan_amnt)) + geom_histogram(aes(fill = loan_status), position = "dodge", bins = 10)+
  labs(x = "Loan amount")
p2<- ggplot(df2, aes(x = installment)) + geom_histogram(aes(fill = loan_status), position = "dodge", bins = 10)+
  labs(x = "Installments")
# Legend function: Fetching legend from ggplot
g_legend<-function(a.gplot){
  tmp <- ggplot_gtable(ggplot_build(a.gplot))
  leg <- which(sapply(tmp$grobs, function(x) x$name) == "guide-box")
  legend <- tmp$grobs[[leg]]
  return(legend)}

mylegend<-g_legend(p1) # use legend of plot 1

# grid arrange : To plot multiple plots on same graph
grid.arrange(
  p1 +theme(legend.position="none"), # remoing individual plot legend
  p2+theme(legend.position="none"),  # remoing individual plot legend
  mylegend,
  nrow = 1,
  top = "Loan amount vs Installments",
  bottom = textGrob(
    "",
    gp = gpar(fontface = 3, fontsize = 9),
    hjust = 1,
    x = 1
  )
)

# however there is a 94% correlation between both Loan amount and installment, its plot does not look very similar
# hence we will keep both for further analysis
```
```{r}
ggplot(df2, aes(x = grade)) + geom_bar(aes(fill = loan_status), position = "dodge")+
  labs(x = "Grades")
Desc(grade ~ loan_status, df2, digits = 1, main = "Grades by Loan Status", plotit = TRUE)
```
```{r}
ggplot(df2, aes(x = home_ownership)) + geom_bar(aes(fill = loan_status), position = "dodge")+
  labs(x = "Home ownership")
```
```{r}
ggplot(df2, aes(x = term)) + geom_bar(aes(fill = loan_status), position = "dodge")+
  labs(x = "Term")
```
```{r}
ggplot(df2, aes(x = purpose)) + geom_bar(aes(fill = loan_status), position = "dodge")+
  labs(x = "Purpose")
```

```{r}
count <- df2 %>%
  group_by(loan_status, grade)%>%
  summarise(n=n())

library('sqldf')
sqlStr <- 
'
SELECT
loan_status, grade, count(*)
from df2
group by loan_status, grade
'
tmp1 <- sqldf(sqlStr)
tmp1


```


```{r}
funded_amnt <-
  df2 %>%
  transmute(loan_amnt = loan_amnt, value = funded_amnt, 
              variable = "funded_amnt")

funded_amnt_inv <-
  df2 %>%
  transmute(loan_amnt = loan_amnt, value = funded_amnt_inv, 
              variable = "funded_amnt_inv")

plot_data <- rbind(funded_amnt, funded_amnt_inv)
```
```{r}
rm(list = ls()[grep("^funded", ls())])
```
```{r}
plot_data %>%
  ggplot(aes(x = loan_amnt, y = value)) +
  facet_wrap(~ variable, scales = "free_x", ncol = 3) +
  geom_point()
```
```{r}
# We can derive a few points from the plot:

# 1. there are instances when funded amount is smaller loan amount
# 2. there seems to be a number of loans where investment is smaller than funded amount i.e. not the full loan is invested in
```


```{r}
# Let's use term in analyzing interest rate by grade
library(scales)
give_count <- 
  stat_summary(fun.data = function(x) return(c(y = median(x)*1.06,
                                               label = length(x))),
               geom = "text")

give_mean <- 
  stat_summary(fun.y = mean, colour = "darkgreen", geom = "point", 
               shape = 18, size = 3, show.legend = FALSE)

df2 %>%
  ggplot(aes(grade, int_rate)) +
  geom_boxplot(fill = "white", colour = "darkblue", 
               outlier.colour = "red", outlier.shape = 1) +
  give_count +
  give_mean +
  scale_y_continuous(labels = comma) +
  labs(title="Interest Rate by Grade", x = "Grade", y = "Interest Rate \n") +
  facet_wrap(~ term)

#interest rate increases with grade increases
#a few loans seem to have an equally low interest rate independent of grade
#the spread of rates seems to increase with grade increase
#there tend to be more outliers on the lower end of the rate
#The 3-year term has a much higher number of high-rated borrowers while the 5-year term has a larger number in the low-rating grades borrowers
```

```{r}
# Let's look into it by subgrades

plot3 <- ggplot(df2, aes(x=int_rate, y=sub_grade)) + geom_point(aes(color=loan_status, alpha=0.4))
plot3 <- plot3 + theme_bw() + scale_fill_manual("Loan Status", values = c("red", "blue")) +
  labs(y="Sub Grades", x="Interest Rates")
plot3

```



```{r}
# Variables to remove:
rm_col <- c("funded_amnt", "funded_amnt_inv", "sub_grade", "emp_title",
    "Taxes.on.production.and.imports.less.subsidies",
    "Gross.domestic.product..GDP..by.state")
df3<- df2[, !(colnames(df2) %in% rm_col), drop = FALSE]
str(df3)
```

```{r}
write.csv(df3, file = "Data1.csv" )
```
