==> 3.2. Feature engineering.Rmd <==
---
title: '3.2 Feature Engineering'
author: "Himanshu"
date: "April 9, 2019"
output: html_document
---

```{r}
library(dplyr)
library(caret)
library(lubridate)
library(stringr)
library(tidyr)
library(corrplot)
library(ggcorrplot)
library(leaps)
library(MASS)
library(ggplot2)
library(standardize)
```

```{r}
df <- read.csv("Data1.csv", header = TRUE)

```

```{r}
str(df)

```

```{r}
#Converting loan status to logical variable TRUE FALSE
df <-
  df %>%
  mutate(default = ifelse(!(loan_status %in% "Defaulted"), FALSE, TRUE))
table(df$default)
df %>%
  summarise(default_freq = sum(default / n()))
table(df$default) / nrow(df)
```

```{r}
df$earliest_cr_line <- as.Date(df$earliest_cr_line)
df$issue_d <- as.Date(df$issue_d)
```

```{r}
# creating region list for address state

NE.name <- c("Connecticut","Maine","Massachusetts","New Hampshire",
             "Rhode Island","Vermont","New Jersey","New York",
             "Pennsylvania")
NE.abrv <- c("CT","ME","MA","NH","RI","VT","NJ","NY","PA")
NE.ref <- c(NE.name,NE.abrv)

MW.name <- c("Indiana","Illinois","Michigan","Ohio","Wisconsin",
             "Iowa","Kansas","Minnesota","Missouri","Nebraska",
             "North Dakota","South Dakota")
MW.abrv <- c("IN","IL","MI","OH","WI","IA","KS","MN","MO","NE",
             "ND","SD")
MW.ref <- c(MW.name,MW.abrv)

S.name <- c("Delaware","District of Columbia","Florida","Georgia",
            "Maryland","North Carolina","South Carolina","Virginia",
            "West Virginia","Alabama","Kentucky","Mississippi",
            "Tennessee","Arkansas","Louisiana","Oklahoma","Texas")
S.abrv <- c("DE","DC","FL","GA","MD","NC","SC","VA","WV","AL",
            "KY","MS","TN","AR","LA","OK","TX")
S.ref <- c(S.name,S.abrv)

W.name <- c("Arizona","Colorado","Idaho","New Mexico","Montana",
            "Utah","Nevada","Wyoming","Alaska","California",
            "Hawaii","Oregon","Washington")
W.abrv <- c("AZ","CO","ID","NM","MT","UT","NV","WY","AK","CA",
            "HI","OR","WA")
W.ref <- c(W.name,W.abrv)

region.list <- list(
  Northeast=NE.ref,
  Midwest=MW.ref,
  South=S.ref,
  West=W.ref)
df$regions <- sapply(df$addr_state, 
                 function(x) names(region.list)[grep(x,region.list)])
df$regions <- as.factor(df$regions)
head(df$regions)
```
```{r}
numVars <- names(which(sapply(df, is.numeric)))
catvars <- names(which(sapply(df, is.factor)))
summary(df[,numVars])
```
```{r}
# creating month column from issue date to see if month has some specific pattern associated with interest rate or loan_status 

df$month <- month(as.POSIXlt(df$issue_d, format="%d/%m/%Y"))
df$month <- as.factor(df$month)
Desc(loan_status ~ month, df, digits = 1, main = "Loan status by month", plotit = TRUE)
Desc(int_rate ~ month, df, digits = 1, main = "Loan status by month", plotit = TRUE)
```

```{r}
# Variables to remove:

rm_col <- c("X","loan_status", "addr_state", "issue_d", "YearBefore","issue_d", "earliest_cr_line")
df2<- df[, !(colnames(df) %in% rm_col), drop = FALSE]
str(df2)
```

```{r}
str(df2)
write.csv(df2, file = "TransfromedData.csv" )
```

```{r}
# We'll do stratified sampling to split our data into training and test sets
# We will use the same test and train set for all models
set.seed(1)
targetVar <- 'default'
inTrain <- createDataPartition(y = df2[,targetVar], list = FALSE, p = .8)
train <- df2[inTrain,]
test <- df2[-inTrain,]
stopifnot(nrow(train) + nrow(test) == nrow(df2))
names(df2)
```

```{r}
write.csv(test, file = "TestData.csv" )
write.csv(train, file = "TrainData.csv" )

```
