==> 3.9 Data Visualization.Rmd <==
---
title: "3.9 Data Visualization"
author: "Vadym Ovcharenko"
date: "4/16/2019"
output: 
  html_document:
    smart: false
---

# Data Visuzalization

In this project, I attempt to execute some of the most interesting data visualization techniques.

```{r}
# dependencies
library(Rtsne)
library(dplyr)
library(splitstackshape)
library(cluster)
library(ggplot2)
library(plotly)
library(purrr)
library(dbscan)
library(caret)
```


## t-SNE

Humans can only percept data in a low-dimensional spaces, such as 2-D or 3-D. Unfortunately, most of the real-world data has many more dimensions. In order to visualize such complex data, we need to somehow map it to a lower dimension space. One of the techniques that is used to do so is called [Dimensionality Reduction](https://en.wikipedia.org/wiki/Dimensionality_reduction). One of the most known dimensionality reduction techniaues is PCA. However, PCA tries to find linear patterns in data and preserve them while going to lower dimensions, treating everything else as random noise. However, it is not always the case that the patterns in the data follow linear fashion. For these more complex data, we need a more advanced dimensionality reduction technique called [t-SNE](https://www.analyticsvidhya.com/blog/2017/01/t-sne-implementation-r-python/) which is almost 100 years younger than PCA. 
Under the hood, t-SNE works by clustering the data using similarity measures. Therefore some data preprocessing is required to ensure that the data is ready for this type of data analysis. Numerical features should be normalized. Categorical features should be leveled in order to calculate the distance between them.

```{r}
df <- read.csv('TransfromedData_1.csv', header = T)
# drop useless columns 
df <- df[, -c(1, 2)]

# application_type is very unbalanced, so we will only explore the individual applications
# also, I didn't want the time dimension to complicate the visualization, so I dropped the date cols.
# additionaly, I wanted the vizualisations to be more demographic, so I dropped variables that are
# characteristics of the loan (except purpose)
df <- select(df[df$application_type == 'INDIVIDUAL',], 
             -c(
               application_type, 
               GeoName, 
               YearBefore,
               regions,
               earliest_cr_line,
               issue_d,
               initial_list_status, 
               Taxes.on.production.and.imports, # drop state-related variables
               Subsidies,
               Real.GDP.by.state,
               Quantity.indexes.for.real.GDP.by.state,
               Per.capita.real.GDP.by.state,
               Gross.operating.surplus,
               Compensation.of.employees
             ))

# make sure that ordinal variables are ordered
df$grade <- factor(df$grade, ordered=T)

catVars <- c('addr_state', 'grade', 'home_ownership', 'verification_status', 'loan_status', 'purpose')
numVars <- setdiff(colnames(df), catVars)

summary(df)
```
I played with a sample of data, because visializing almost a million rows is quite challanging computationally. 

```{r}

set.seed(1)
# Take a stratified sample of data
dfs <- stratified(df, c("loan_status"), 0.01)

write.csv(space, file = "space_3d.csv", row.names=FALSE)

summary(dfs)
colnames(dfs)
#sapply(catVars, function(x) plot(dfs[, ..x], main=x))
```


```{r}
# calculate distance for variables of all types
gower_dist <- dfs %>% select(-loan_status) %>% daisy(
                    metric = "gower")
summary(gower_dist)
```


```{r}
tsne_obj <- Rtsne(gower_dist, is_distance = TRUE, perplexity=50)

tsne_data <- tsne_obj$Y %>%
  data.frame() %>%
  setNames(c("x_coord", "y_coord")) %>%
  cbind(dfs)

plt <- ggplot(aes(x = x_coord, y = y_coord), data = tsne_data) + 
  geom_point(aes(color = loan_status, shape=purpose), alpha=0.5)
plt
```
Unfortunately, t-SNE does not show any apparent spearation of defaulters. Although it shows us that there are some very distinct clusters in our dataset. Looking closer into the clusters, shows that they are almost perfectly based on the state variable.

```{r}
plt + geom_text(aes(label=addr_state), check_overlap = T)
```
Obviously, t-SNE algorithm tells us that state is a major distinguishing factor between groups of customers in our dataset. 

```{r}
# calculate distance for variables of all types
gower_dist <- dfs %>% select(-c(loan_status, addr_state)) %>% daisy(
                    metric = "gower")

tsne_obj <- Rtsne(gower_dist, is_distance = TRUE, perplexity=50)

tsne_data <- tsne_obj$Y %>%
  data.frame() %>%
  setNames(c("x_coord", "y_coord")) %>%
  cbind(dfs)

ggplot(aes(x = x_coord, y = y_coord), data = tsne_data) + 
  geom_point(aes(color = loan_status, shape=purpose), alpha=0.5)+
  geom_text(aes(label=addr_state), check_overlap = T)
```
Removing the state variable adds more personality to the vizualisation and makes it more interesting to explore.

```{r}


plot_ly(tsne_data, x = ~x_coord, y = ~y_coord, 
        color = ~loan_status, 
        alpha=0.5)
```
```{r}
gower_dist <- dfs %>% select(-c(loan_status, addr_state)) %>% daisy(
                    metric = "gower")
tsne_obj <- Rtsne(gower_dist, is_distance = TRUE, perplexity=50, dims=3)

tsne_data <- tsne_obj$Y %>%
  data.frame() %>%
  setNames(c("x_coord", "y_coord", "z_coord")) 

plot_ly(tsne_data %>% cbind(dfs)
        , x = ~x_coord, y = ~y_coord, z = ~z_coord, 
        color = ~loan_status, 
        shape = ~purpose,
        mode='markers',
        hoverinfo = 'text',
        text = ~paste('Purpose', dfs$purpose) # TODO display all relevant data
        )
```

```{r}

# validate t-SNE results
set.seed(1)
trainIndex <- createDataPartition(y = dfs$loan_status, list=F, p = 0.8)

opt <- optics(tsne_data, minPts = 50)
clust <- extractXi(opt, xi = 0.02)

# mark test rows
dfs$test <- F
dfs$test[-trainIndex] <- T

space <- tsne_data %>% cbind(dfs, cluster = clust$cluster) %>%
  group_by(cluster) %>%
  mutate(cluster_default_rate = 
           ifelse(sum(test == F) != 0, 
                  sum(loan_status == 'Defaulted' && test == F) / sum(test == F), 
                  0)) %>%
  ungroup()

threshold <- 0.015
predicted <- space[-trainIndex, 'cluster_default_rate'] >= threshold
true <- space[-trainIndex, 'loan_status'] == 'Defaulted'
confusionMatrix(predicted, true, positive = 'TRUE')


p <- plot_ly() %>% add_trace(x = space$x_coord, y = space$y_coord, z = space$z_coord, 
        mode='markers',
        hoverinfo = 'text',
        marker=list(
          color = space$cluster_default_rate, 
          colorscale='RdYlBu',
          colorbar=list(title='Default rate')
        ),
        text = ~paste('</br>Cluster', clust$cluster, 
                      '</br>Cluster default rate:', space$cluster_default_rate, 
                      '</br>Purpose', space$purpose), # TODO display all relevant data,
        type = "scatter3d"
        ) 
p

new_data_point <- space[space$test == T, ][5,]

p %>% layout(scene = list(
      aspectratio = list(
        x = 1,
        y = 1,
        z = 1
      ),
      camera = list(
        center = list(
          x = 0,
          y = 0,
          z = 0
        ),
        eye = list(
          x = 1.96903462608,
          y = -1.09022831971,
          z = 0.405345349304
        ),
        up = list(
          x = 0,
          y = 0,
          z = 1
        )
      ), annotations = list(list(
        showarrow = T,
        # datapoint coordinates:
        x = new_data_point$x_coord,
        y = new_data_point$y_coord,
        z = new_data_point$z_coord,
        text = "New loan application",
        xshift = 0,
        opacity = 1
      ))))
```

```{r}
plot_ly() %>%
  add_trace(
    x = c("2017-01-01", "2017-02-10", "2017-03-20"), 
    y = c("A", "B", "C"), 
    z = c(1, 1000, 100000), 
    name = "z", 
    type = "scatter3d"
  ) %>% 
  layout(
    scene = list(
      aspectratio = list(
        x = 1,
        y = 1,
        z = 1
      ),
      camera = list(
        center = list(
          x = 0,
          y = 0,
          z = 0
        ),
        eye = list(
          x = 1.96903462608,
          y = -1.09022831971,
          z = 0.405345349304
        ),
        up = list(
          x = 0,
          y = 0,
          z = 1
        )
      ),
      dragmode = "turntable",
      xaxis = list(
        title = "",
        type = "date"
      ),
      yaxis = list(
        title = "",
        type = "category"
      ),
      zaxis = list(
        title = "",
        type = "log"
      ),
      annotations = list(list(
        showarrow = F,
        x = "2017-01-01",
        y = "A",
        z = 0,
        text = "Point 1",
        xanchor = "left",
        xshift = 10,
        opacity = 0.7
      ), list(
        x = "2017-02-10",
        y = "B",
        z = 4,
        text = "Point 2",
        textangle = 0,
        ax = 0,
        ay = -75,
        font = list(
          color = "black",
          size = 12
        ),
        arrowcolor = "black",
        arrowsize = 3,
        arrowwidth = 1,
        arrowhead = 1
      ), list(
        x = "2017-03-20",
        y = "C",
        z = 5,
        ax = 50,
        ay = 0,
        text = "Point 3",
        arrowhead = 1,
        xanchor = "left",
        yanchor = "bottom"
      )
    )),
    xaxis = list(title = "x"),
    yaxis = list(title = "y")
  )

```
