==> 4.1 results rShiny.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

```{r}
library(DescTools)
library(ggplot2)
library(dplyr)
library(caret)
library(lubridate)
library(stringr)
library(tidyr)
library(corrplot)
library(ggcorrplot)
library(leaps)
library(MASS)
library(ggplot2)
library(standardize)
library(Amelia)
library(scales)
library(ROCR)

log <- read.csv("C:/Users/young/Downloads/Project-Files-master/Project-Files-master/LogisticTest/LogisticTest.csv")
head(log)

# Compute for actual profit
profit_actual <- sum(log$profit)

# Compute for profit based on log model
profit_log <- sum(log$profit * (!log$model_pred_t))
profit_log

# Compute for yield based on actual profit
yield_actual <- sum(log$profit)/sum(log$loan_amnt)
yield_actual
 
# Copute for yield based on log model
yield_log <- sum(log$profit * (!log$model_pred_t))/sum(log$loan_amnt * (!log$model_pred_t))
yield_log

tsne <- read.csv("C:/Users/young/Downloads/Project-Files-master/Project-Files-master/tsne_predictions.csv")
tsne <- merge(tsne, log, by = "X")
head(tsne)

# Compute for profit based on tsne
profit_tsne <- sum((tsne$cluster_default_rate < 0.001) * tsne$profit)
profit_tsne

# Compute for yield based on tsne
yield_tsne <- sum((tsne$cluster_default_rate < 0.001) * tsne$profit)/sum(tsne$loan_amnt * (tsne$cluster_default_rate < 0.001))
yield_tsne

tree <- read.csv("C:/Users/young/Downloads/Project-Files-master/Project-Files-master/DecisionTreeTest.csv")
head(tree)
# Compute for profit based on tree
profit_tree <- sum((!tree$fit.pred) * tree$profit)
profit_tree

# Compute for yield based on tree
yield_tree <- sum((!tree$fit.pred) * tree$profit)/sum(tree$loan_amnt * (!tree$fit.pred))
yield_tree

xgb_mono <- read.csv("C:/Users/young/Downloads/Project-Files-master/Project-Files-master/pred/XGB_mono_pred.csv")
xgb_mono <- merge(xgb_mono, log, by = "X")

summary(xgb_mono$loan_amnt)

# Compute for profit based on tree
profit_xgb <- sum((!xgb_mono$XGB_mono_pred) * xgb_mono$profit)
profit_xgb

# Compute for yield based on tree
yield_xgb <- sum((!xgb_mono$XGB_mono_pred) * xgb_mono$profit)/sum((!xgb_mono$XGB_mono_pred) * xgb_mono$loan_amnt)
yield_xgb

# XNN
xnn <- read.csv("C:/Users/young/Downloads/Project-Files-master/Project-Files-master/pred/XNN_pred.csv")
xnn <- merge(xnn, log, by = "X")
head(xnn)

sum(xnn$XNN_pred * log$model_pred_t)

# Compute for profit based on tree
profit_xnn <- sum((!xnn$XNN_pred) * xnn$profit)
profit_xnn

# Compute for yield based on tree
yield_xnn <- sum((!xnn$XNN_pred) * xnn$profit)/sum((!xnn$XNN_pred) * xnn$loan_amnt)
yield_xnn


train <- read.csv("C:/Users/young/Downloads/Project-Files-master/Project-Files-master/TrainTestData/TrainData.csv")  
vars_to_remove <- c("X.1", "X","pymnt_plan", "GeoName", "application_type")
train <- train[, ! names(train) %in% vars_to_remove, drop = F] 
levels(train$term) <- c("36 months", "60 months")

log_model <- glm(default ~ term+home_ownership+purpose+total_acc+int_rate+annual_inc+ dti+regions
                     , family=binomial(link='logit'), data=train)
```
```{r}
library(ggplot2)
library(Cairo)   # For nicer ggplot2 output when deployed on Linux
# dependencies
require('Rtsne')
library(dplyr)
library(splitstackshape)
library(cluster)
library(ggplot2)
library(plotly)


tsne_data <- read.csv('C:/Users/young/Downloads/Project-Files-master/Project-Files-master/space_2d.csv', header = T)
space <- read.csv('C:/Users/young/Downloads/Project-Files-master/Project-Files-master/space_3d.csv', header = T)

```


```{r}
library(shiny)
library(shinydashboard)
library(scales)

# Add pie chart of loan vs yielded profits

ui <- dashboardPage(
  dashboardHeader(title = "Lending Club Loan Dashboard"),
  
  dashboardSidebar(
    sidebarMenu(
      menuItem("Summary", tabName = "summary", icon = icon("dashboard")),
      menuItem("TSNE", tabName = "eda", icon = icon("dashboard")),
      # menuItem("Modelling", tabName = "modelling", icon = icon("dashboard")),
      menuItem("New Applicant", tabName = "new_applicant", icon = icon("dashboard"))
    )
  ),
  
  dashboardBody(
     tabItems(
       
      # First tab content
      tabItem(tabName = "summary",
        fluidRow(
          # infoBoxes with fill=TRUE
          tabBox(title = "Model Used",
                 id = "tabsetSummary",
                 height = "1000px",
                 width = "1000px",
                 tabPanel("Logistic",
                          
                          box(infoBoxOutput("statusQuo_log"),
                              infoBoxOutput("amountSaved_log")),
                          box(infoBoxOutput("yield_actual_log"),
                              infoBoxOutput("yield_log")),
                          tags$style("#statusQuo_log {width:400px;}"),
                          tags$style("#amountSaved_log {width:400px;}"),
                          tags$style("#yield_actual_log {width:400px;}"),
                          tags$style("#yield_log {width:400px;}"),
                          tabBox(title = "By Term",
                                 id = "tab_Term_Log",
                                 tabPanel("Frequency",
                                          box(plotOutput("defaults_by_term_log", height = 400, width = 600))
                                          ),
                                 tabPanel("Amount",
                                          box(plotOutput("defaults_by_term_amount_log", height = 400, width = 600))
                                          )
                                ),
                          tabBox(title = "By Grade",
                                 id = "tab_Grade",
                                 tabPanel("Frequency",
                                          box(plotOutput("defaults_by_grade_log", height = 400, width = 600))
                                          ),
                                 tabPanel("Amount",
                                          box(plotOutput("defaults_by_grade_amount_log", height = 400, width = 600))
                                          )
                                 )
                           ),
                 
                 tabPanel("Decision Tree",
                          box(infoBoxOutput("statusQuo_tree"),
                              infoBoxOutput("amountSaved_tree")),
                          box(infoBoxOutput("yield_actual_tree"),
                              infoBoxOutput("yield_tree")),
                          tags$style("#statusQuo_tree {width:400px;}"),
                          tags$style("#amountSaved_tree {width:400px;}"),
                          tags$style("#yield_actual_tree {width:400px;}"),
                          tags$style("#yield_tree {width:400px;}"),
                          tabBox(title = "By Term",
                                 id = "tab_Term_Tree",
                                 tabPanel("Frequency",
                                          box(plotOutput("defaults_by_term_tree", height = 400, width = 600))
                                          ),
                                 tabPanel("Amount",
                                          box(plotOutput("defaults_by_term_amount_tree", height = 400, width = 600))
                                          )
                                ),
                          tabBox(title = "By Grade",
                                 id = "tab_Grade_Tree",
                                 tabPanel("Frequency",
                                          box(plotOutput("defaults_by_grade_tree", height = 400, width = 600))
                                          ),
                                 tabPanel("Amount",
                                          box(plotOutput("defaults_by_grade_amount_tree", height = 400, width = 600))
                                          )
                                 )
                          ),
                 tabPanel("TSNE",
                          box(infoBoxOutput("statusQuo_tsne"),
                              infoBoxOutput("amountSaved_tsne")),
                          box(infoBoxOutput("yield_actual_tsne"),
                              infoBoxOutput("yield_tsne")),
                          tags$style("#statusQuo_tsne {width:400px;}"),
                          tags$style("#amountSaved_tsne {width:400px;}"),
                          tags$style("#yield_actual_tsne {width:400px;}"),
                          tags$style("#yield_tsne {width:400px;}"),
                          tabBox(title = "By Term",
                                 id = "tab_Term_TSNE",
                                 tabPanel("Frequency",
                                          box(plotOutput("defaults_by_term_tsne", height = 400, width = 600))
                                          ),
                                 tabPanel("Amount",
                                          box(plotOutput("defaults_by_term_amount_tsne", height = 400, width = 600))
                                          )
                                ),
                          tabBox(title = "By Grade",
                                 id = "tab_Grade_TSNE",
                                 tabPanel("Frequency",
                                          box(plotOutput("defaults_by_grade_tsne", height = 400, width = 600))
                                          ),
                                 tabPanel("Amount",
                                          box(plotOutput("defaults_by_grade_amount_tsne", height = 400, width = 600))
                                          )
                                 )
                          ),
          
                tabPanel("XGB",
                          box(infoBoxOutput("statusQuo_xgb"),
                              infoBoxOutput("amountSaved_xgb")),
                          box(infoBoxOutput("yield_actual_xgb"),
                              infoBoxOutput("yield_xgb")),
                          tags$style("#statusQuo_xgb {width:400px;}"),
                          tags$style("#amountSaved_xgb {width:400px;}"),
                          tags$style("#yield_actual_xgb {width:400px;}"),
                          tags$style("#yield_xgb {width:400px;}"),
                          tabBox(title = "By Term",
                                 id = "tab_Term_xgb",
                                 tabPanel("Frequency",
                                          box(plotOutput("defaults_by_term_xgb", height = 400, width = 600))
                                          ),
                                 tabPanel("Amount",
                                          box(plotOutput("defaults_by_term_amount_xgb", height = 400, width = 600))
                                          )
                                ),
                          tabBox(title = "By Grade",
                                 id = "tab_Grade_xgb",
                                 tabPanel("Frequency",
                                          box(plotOutput("defaults_by_grade_xgb", height = 400, width = 600))
                                          ),
                                 tabPanel("Amount",
                                          box(plotOutput("defaults_by_grade_amount_xgb", height = 400, width = 600))
                                          )
                                 )
                          ),
        
                tabPanel("XNN",
                          box(infoBoxOutput("statusQuo_xnn"),
                              infoBoxOutput("amountSaved_xnn")),
                          box(infoBoxOutput("yield_actual_xnn"),
                              infoBoxOutput("yield_xnn")),
                          tags$style("#statusQuo_xnn {width:400px;}"),
                          tags$style("#amountSaved_xnn {width:400px;}"),
                          tags$style("#yield_actual_xnn {width:400px;}"),
                          tags$style("#yield_xnn {width:400px;}"),
                          tabBox(title = "By Term",
                                 id = "tab_Term_xnn",
                                 tabPanel("Frequency",
                                          box(plotOutput("defaults_by_term_xnn", height = 400, width = 600))
                                          ),
                                 tabPanel("Amount",
                                          box(plotOutput("defaults_by_term_amount_xnn", height = 400, width = 600))
                                          )
                                ),
                          tabBox(title = "By Grade",
                                 id = "tab_Grade_xnn",
                                 tabPanel("Frequency",
                                          box(plotOutput("defaults_by_grade_xnn", height = 400, width = 600))
                                          ),
                                 tabPanel("Amount",
                                          box(plotOutput("defaults_by_grade_amount_xnn", height = 400, width = 600))
                                          )
                                 )
                          )
                 )
        )
      ),

      # Second tab content
      tabItem(tabName = "eda",
        fluidRow(
          column(width = 12,
          plotlyOutput("plot2", height = 600
                      # Equivalent to: click = clickOpts(id = "plot_click")
                      #click = "plot1_click",
                      #brush = brushOpts(
                      #  id = "plot1_brush"
                      #)
                      )
                )
          ),
          fluidRow(
          column(width = 12,
          plotOutput("plot1", height = 600,
                                # Equivalent to: click = clickOpts(id = "plot_click")
                                click = "plot1_click",
                                brush = brushOpts(
                                  id = "plot1_brush"
                                )
                    )
                )
          ),
            fluidRow(
            column(width = 12,
                     h4("Brushed points"),
                     dataTableOutput("brush_info")
                    )
          )
      ),
      
      # # third tab content
      # tabItem(tabName = "modelling"
      # ),
      
      # Fourth tab content
      tabItem(tabName = "new_applicant",
        fluidRow(
          # box(selectInput("model", "Select Model", c("Logistic", "Decision Tree", "Random Forest"))),
          
          tabBox(title = "Model Used",
                 id = "tabset1",
                 height = "1000px",
                 width = "1000px",
                  tabPanel("Logistic", 
                           box(textInput("text1", "Enter Term")),
                           box(textInput("text2", "Enter Home Ownership")),
                           box(textInput("text3", "Enter Purpose")),
                           box(textInput("text4", "Enter Total Account")),
                           box(textInput("text5", "Enter Interest Rate")),
                           box(textInput("text6", "Enter Annual Income")),
                           box(textInput("text7", "Enter DTI")),
                           box(textInput("text8", "Enter Region")),
                           #(term, home_ownership, purpose, total_acc, int_rate, annual_inc, dti, regions
                           box(actionButton("log_button", "Predict")),
                           box(verbatimTextOutput("log_class"))
                  ),
                  tabPanel("Decision Tree",
                           box(textInput("text9", "Enter Term")),
                           box(textInput("text10", "Enter Public Record")),
                           box(textInput("text11", "Enter Initial List Status")),
                           box(textInput("text12", "Enter Quantity Index for Real GDP by State")),
                           box(textInput("text13", "Enter Interest Rate")),
                           box(textInput("text14", "Enter Annual Income")),
                           box(textInput("text15", "Enter Real GDP by State")),
                           box(textInput("text16", "Enter Month")),
                           
                           box(actionButton("tree_button", "Predict")),
                           box(verbatimTextOutput("tree_class"))
                  ),
                  tabPanel("TSNE",
                          box(textInput("text17", "Enter Term")),
                           box(textInput("text18", "Enter Public Record")),
                           box(textInput("text19", "Enter Initial List Status")),
                           box(textInput("text20", "Enter Quantity Index for Real GDP by State")),
                           box(textInput("text21", "Enter Interest Rate")),
                           box(textInput("text22", "Enter Annual Income")),
                           box(textInput("text23", "Enter Real GDP by State")),
                           box(textInput("text24", "Enter Month")),
                           
                           box(actionButton("tsne_button", "Predict")),
                           box(verbatimTextOutput("tsne_class"))
                  ),
                  tabPanel("XGB Mono",
                          box(textInput("text25", "Enter Term")),
                           box(textInput("text26", "Enter Public Record")),
                           box(textInput("text27", "Enter Initial List Status")),
                           box(textInput("text28", "Enter Quantity Index for Real GDP by State")),
                           box(textInput("text29", "Enter Interest Rate")),
                           box(textInput("text30", "Enter Annual Income")),
                           box(textInput("text31", "Enter Real GDP by State")),
                           box(textInput("text32", "Enter Month")),
                           
                           box(actionButton("xgb_button", "Predict")),
                           box(verbatimTextOutput("xgb_class"))
                  ),
                  tabPanel("XNN",
                          box(textInput("text33", "Enter Term")),
                           box(textInput("text34", "Enter Public Record")),
                           box(textInput("text35", "Enter Initial List Status")),
                           box(textInput("text36", "Enter Quantity Index for Real GDP by State")),
                           box(textInput("text37", "Enter Interest Rate")),
                           box(textInput("text38", "Enter Annual Income")),
                           box(textInput("text39", "Enter Real GDP by State")),
                           box(textInput("text40", "Enter Month")),
                           
                           box(actionButton("xnn", "Predict")),
                           box(verbatimTextOutput("xnn_class"))
                  )
          )
          
          # textInput("text2", "Enter text2"),
          # textInput("text3", "Enter text3"),
          # textInput("text4", "Enter text4"),
          # textInput("text5", "Enter text5")
          # 
        )
      )
    )
  )
)

server <- function(input, output, session) {
  
  # LOGISTIC
  
  output$defaults_by_term_log <- renderPlot({
    ggplot(log, aes(x = term, fill = model_pred_t)) + geom_bar() + ggtitle("Number of Defaults by Term") # Add the percentage of correctly predicted defaults to emphasize the effect
  })
  
  output$defaults_by_term_amount_log <- renderPlot({
    # Rather than count, try computing for the amount that could have been saved by Grade
    ggplot(log, aes(x = term, fill = model_pred_t, y = loan_amnt)) + geom_col() + scale_y_continuous(labels = dollar)+ 
      ggtitle("Total Dollar Amount of Charged-Off by Term")  #Add the amount saved from correctly predicted defaults 
  })
  
  output$defaults_by_grade_log <- renderPlot({
    ggplot(log, aes(x = grade, fill = model_pred_t)) + geom_bar() + ggtitle("Number of Defaults by Grade") # Add the percentage of correctly predicted defaults to emphasize the effect
  })
  
  output$defaults_by_grade_amount_log <- renderPlot({
    # Rather than count, try computing for the amount that could have been saved by Grade
    ggplot(log, aes(x = grade, fill = model_pred_t, y = loan_amnt)) + geom_col() + scale_y_continuous(labels = dollar)+ 
      ggtitle("Total Dollar Amount of Charged-Off by Grade")  #Add the amount saved from correctly predicted defaults 
  })

  output$statusQuo_log <- renderInfoBox({
    infoBox(
      "Status Quo", dollar(round(profit_actual,0)) , icon = icon("thumbs-up", lib = "glyphicon"),
      color = "green", fill = TRUE
    )
  })
  
  output$amountSaved_log <- renderInfoBox({
    infoBox(
      "Amount Saved from Logistic", dollar(round(profit_log - profit_actual,0)) , icon = icon("thumbs-up", lib = "glyphicon"),
      color = "green", fill = TRUE
    )
  })

  output$yield_actual_log <- renderInfoBox({
    infoBox(
      "Status Quo Yield", round(yield_actual,4) , icon = icon("thumbs-up", lib = "glyphicon"),
      color = "green", fill = TRUE
    )
  })
  
  output$yield_log <- renderInfoBox({
    infoBox(
      "Yield from Logistic", round(yield_log,4) , icon = icon("thumbs-up", lib = "glyphicon"),
      color = "green", fill = TRUE
    )
  })
  

  # DECISION TREE
  output$defaults_by_term_tree <- renderPlot({
    ggplot(tree, aes(x = term, fill = fit.pred)) + geom_bar() + ggtitle("Number of Defaults by Term") # Add the percentage of correctly predicted defaults to emphasize the effect
  })

  output$defaults_by_term_amount_tree <- renderPlot({
    # Rather than count, try computing for the amount that could have been saved by Grade
    ggplot(tree, aes(x = term, fill = fit.pred, y = loan_amnt)) + geom_col() + scale_y_continuous(labels = dollar)+
      ggtitle("Total Dollar Amount of Charged-Off by Term")  #Add the amount saved from correctly predicted defaults
  })

  output$defaults_by_grade_tree <- renderPlot({
    ggplot(tree, aes(x = grade, fill = fit.pred)) + geom_bar() + ggtitle("Number of Defaults by Grade") # Add the percentage of correctly predicted defaults to emphasize the effect
  })

  output$defaults_by_grade_amount_tree <- renderPlot({
    # Rather than count, try computing for the amount that could have been saved by Grade
    ggplot(tree, aes(x = grade, fill = fit.pred, y = loan_amnt)) + geom_col() + scale_y_continuous(labels = dollar)+
      ggtitle("Total Dollar Amount of Charged-Off by Grade")  #Add the amount saved from correctly predicted defaults
  })
  
  output$statusQuo_tree <- renderInfoBox({
    infoBox(
      "Status Quo", dollar(round(profit_actual,0)) , icon = icon("thumbs-up", lib = "glyphicon"),
      color = "green", fill = TRUE
    )
  })
  

  output$amountSaved_tree <- renderInfoBox({
    infoBox(
      "Amount Saved from Decision Tree", dollar(round(profit_tree - profit_actual,0)) , icon = icon("thumbs-up", lib = "glyphicon"),
      color = "green", fill = TRUE
    )
  })
  
   output$yield_actual_tree <- renderInfoBox({
    infoBox(
      "Status Quo Yield", round(yield_actual,4) , icon = icon("thumbs-up", lib = "glyphicon"),
      color = "green", fill = TRUE
    )
  })

  output$yield_tree <- renderInfoBox({
    infoBox(
      "Yield from Decision Tree", round(yield_tree,4) , icon = icon("thumbs-up", lib = "glyphicon"),
      color = "green", fill = TRUE
    )
  })


  # TSNE

  output$defaults_by_term_tsne <- renderPlot({
    ggplot(tsne, aes(x = term, fill = model_pred_t)) + geom_bar() + ggtitle("Number of Defaults by Term") # Add the percentage of correctly predicted defaults to emphasize the effect
  })

  output$defaults_by_term_amount_tsne <- renderPlot({
    # Rather than count, try computing for the amount that could have been saved by Grade
    ggplot(tsne, aes(x = term, fill = model_pred_t, y = loan_amnt)) + geom_col() + scale_y_continuous(labels = dollar)+
      ggtitle("Total Dollar Amount of Charged-Off by Term")  #Add the amount saved from correctly predicted defaults
  })

  output$defaults_by_grade_tsne <- renderPlot({
    ggplot(tsne, aes(x = grade, fill = model_pred_t)) + geom_bar() + ggtitle("Number of Defaults by Grade") # Add the percentage of correctly predicted defaults to emphasize the effect
  })

  output$defaults_by_grade_amount_tsne <- renderPlot({
    # Rather than count, try computing for the amount that could have been saved by Grade
    ggplot(tsne, aes(x = grade, fill = model_pred_t, y = loan_amnt)) + geom_col() + scale_y_continuous(labels = dollar)+
      ggtitle("Total Dollar Amount of Charged-Off by Grade")  #Add the amount saved from correctly predicted defaults
  })

   output$statusQuo_tsne <- renderInfoBox({
    infoBox(
      "Status Quo", dollar(round(profit_actual,0)) , icon = icon("thumbs-up", lib = "glyphicon"),
      color = "green", fill = TRUE
    )
  })

  output$amountSaved_tsne <- renderInfoBox({
    infoBox(
      "Amount Saved from TSNE", dollar(round(profit_tsne - profit_actual,0)) , icon = icon("thumbs-up", lib = "glyphicon"),
      color = "green", fill = TRUE
    )
  })
  
   output$yield_actual_tsne <- renderInfoBox({
    infoBox(
      "Status Quo Yield", round(yield_actual,4) , icon = icon("thumbs-up", lib = "glyphicon"),
      color = "green", fill = TRUE
    )
  })

  output$yield_tsne <- renderInfoBox({
    infoBox(
      "Yield from TSNE", round(yield_tsne,4) , icon = icon("thumbs-up", lib = "glyphicon"),
      color = "green", fill = TRUE
    )
  })
  
   # XGB
  output$defaults_by_term_xgb <- renderPlot({
    ggplot(xgb_mono, aes(x = term, fill = XGB_mono_pred )) + geom_bar() + ggtitle("Number of Defaults by Term") # Add the percentage of correctly predicted defaults to emphasize the effect
  })

  output$defaults_by_term_amount_xgb <- renderPlot({
    # Rather than count, try computing for the amount that could have been saved by Grade
    ggplot(xgb_mono, aes(x = term, fill = XGB_mono_pred, y = loan_amnt)) + geom_col() + scale_y_continuous(labels = dollar)+
      ggtitle("Total Dollar Amount of Charged-Off by Term")  #Add the amount saved from correctly predicted defaults
  })

  output$defaults_by_grade_xgb <- renderPlot({
    ggplot(xgb_mono, aes(x = grade, fill = XGB_mono_pred)) + geom_bar() + ggtitle("Number of Defaults by Grade") # Add the percentage of correctly predicted defaults to emphasize the effect
  })

  output$defaults_by_grade_amount_xgb <- renderPlot({
    # Rather than count, try computing for the amount that could have been saved by Grade
    ggplot(xgb_mono, aes(x = grade, fill = XGB_mono_pred, y = loan_amnt)) + geom_col() + scale_y_continuous(labels = dollar)+
      ggtitle("Total Dollar Amount of Charged-Off by Grade")  #Add the amount saved from correctly predicted defaults
  })

   output$statusQuo_xgb <- renderInfoBox({
    infoBox(
      "Status Quo", dollar(round(profit_actual,0)) , icon = icon("thumbs-up", lib = "glyphicon"),
      color = "green", fill = TRUE
    )
  })

  output$amountSaved_xgb <- renderInfoBox({
    infoBox(
      "Amount Saved from XGB", dollar(round(profit_xgb - profit_actual,0)) , icon = icon("thumbs-up", lib = "glyphicon"),
      color = "green", fill = TRUE
    )
  })
  
   output$yield_actual_xgb <- renderInfoBox({
    infoBox(
      "Status Quo Yield", round(yield_actual,4) , icon = icon("thumbs-up", lib = "glyphicon"),
      color = "green", fill = TRUE
    )
  })

  output$yield_xgb <- renderInfoBox({
    infoBox(
      "Yield from XGB", round(yield_xgb,4) , icon = icon("thumbs-up", lib = "glyphicon"),
      color = "green", fill = TRUE
    )
  })
  
   # XNN
  
  output$defaults_by_term_xnn <- renderPlot({
    ggplot(xnn, aes(x = term, fill = XNN_pred )) + geom_bar() + ggtitle("Number of Defaults by Term") # Add the percentage of correctly predicted defaults to emphasize the effect
  })

  output$defaults_by_term_amount_xnn <- renderPlot({
    # Rather than count, try computing for the amount that could have been saved by Grade
    ggplot(xnn, aes(x = term, fill = XNN_pred, y = loan_amnt)) + geom_col() + scale_y_continuous(labels = dollar)+
      ggtitle("Total Dollar Amount of Charged-Off by Term")  #Add the amount saved from correctly predicted defaults
  })

  output$defaults_by_grade_xnn <- renderPlot({
    ggplot(xnn, aes(x = grade, fill = XNN_pred)) + geom_bar() + ggtitle("Number of Defaults by Grade") # Add the percentage of correctly predicted defaults to emphasize the effect
  })

  output$defaults_by_grade_amount_xnn <- renderPlot({
    # Rather than count, try computing for the amount that could have been saved by Grade
    ggplot(xnn, aes(x = grade, fill = XNN_pred, y = loan_amnt)) + geom_col() + scale_y_continuous(labels = dollar)+
      ggtitle("Total Dollar Amount of Charged-Off by Grade")  #Add the amount saved from correctly predicted defaults
  })

   output$statusQuo_xnn <- renderInfoBox({
    infoBox(
      "Status Quo", dollar(round(profit_actual,0)) , icon = icon("thumbs-up", lib = "glyphicon"),
      color = "green", fill = TRUE
    )
  })

  output$amountSaved_xnn <- renderInfoBox({
    infoBox(
      "Amount Saved from XGB", dollar(round(profit_xnn - profit_actual,0)) , icon = icon("thumbs-up", lib = "glyphicon"),
      color = "green", fill = TRUE
    )
  })
  
   output$yield_actual_xnn <- renderInfoBox({
    infoBox(
      "Status Quo Yield", round(yield_actual,4) , icon = icon("thumbs-up", lib = "glyphicon"),
      color = "green", fill = TRUE
    )
  })

  output$yield_xnn <- renderInfoBox({
    infoBox(
      "Yield from XGB", round(yield_xnn,4) , icon = icon("thumbs-up", lib = "glyphicon"),
      color = "green", fill = TRUE
    )
  })
  
  bb <- ifelse(predict(log_model, train, type = "response") >= 0.5, "Default", "Not Default")
  train[194,] %>% select(term, home_ownership, purpose, total_acc, int_rate, annual_inc, dti, regions)
  
  # Prediction
  observeEvent(input$log_button, {
    to_be_predicted <- data.frame(
      term = input$text1, 
      home_ownership = input$text2, 
      purpose = input$text3, 
      total_acc = as.integer(input$text4), 
      int_rate = as.numeric(input$text5), 
      annual_inc = as.numeric(input$text6), 
      dti = as.numeric(input$text7), 
      regions = input$text8)
    
    prob <- predict(log_model, to_be_predicted, type = "response")
    class <- ifelse(prob >= 0.5, "Default", "Not Default")
    output$log_class <- renderText({class})
  })
  
  observeEvent(input$tree_button, {
    to_be_predicted <- data.frame(
      term = input$text1, 
      pub_rec = as.integer(input$text2), 
      initial_list_status = input$text3, 
      Quantity.indexes.for.real.GDP.by.state = as.integer(input$text4), 
      int_rate = as.numeric(input$text5), 
      annual_inc = as.numeric(input$text6), 
      Real.GDP.by.state = as.integer(input$text7), 
      month = as.integer(input$text8))
    
    prob <- predict(final_model, to_be_predicted, type = "response")
    class <- ifelse(prob >= 0.5, "Not Default", "Default")
    output$tree_class <- renderText({class})
  })
  
  observeEvent(input$tsne_button, {
    to_be_predicted <- data.frame(
      term = input$text1, 
      pub_rec = as.integer(input$text2), 
      initial_list_status = input$text3, 
      Quantity.indexes.for.real.GDP.by.state = as.integer(input$text4), 
      int_rate = as.numeric(input$text5), 
      annual_inc = as.numeric(input$text6), 
      Real.GDP.by.state = as.integer(input$text7), 
      month = as.integer(input$text8))
    
    prob <- predict(final_model, to_be_predicted, type = "response")
    class <- ifelse(prob >= 0.5, "Not Default", "Default")
    output$tsne_class <- renderText({class})
  })
  
  observeEvent(input$xgb_button, {
    to_be_predicted <- data.frame(
      term = input$text1, 
      pub_rec = as.integer(input$text2), 
      initial_list_status = input$text3, 
      Quantity.indexes.for.real.GDP.by.state = as.integer(input$text4), 
      int_rate = as.numeric(input$text5), 
      annual_inc = as.numeric(input$text6), 
      Real.GDP.by.state = as.integer(input$text7), 
      month = as.integer(input$text8))
    
    prob <- predict(final_model, to_be_predicted, type = "response")
    class <- ifelse(prob >= 0.5, "Not Default", "Default")
    output$xgb_class <- renderText({class})
  })
  
  observeEvent(input$xnn_button, {
    to_be_predicted <- data.frame(
      term = input$text1, 
      pub_rec = as.integer(input$text2), 
      initial_list_status = input$text3, 
      Quantity.indexes.for.real.GDP.by.state = as.integer(input$text4), 
      int_rate = as.numeric(input$text5), 
      annual_inc = as.numeric(input$text6), 
      Real.GDP.by.state = as.integer(input$text7), 
      month = as.integer(input$text8))
    
    prob <- predict(final_model, to_be_predicted, type = "response")
    class <- ifelse(prob >= 0.5, "Not Default", "Default")
    output$xnn_class <- renderText({class})
  })
  
  output$plot1 <- renderPlot({
    ggplot(aes(x = x_coord, y = y_coord), data = tsne_data) + 
      geom_point(aes(color = loan_status, shape=purpose), alpha=0.5)
  })
  
  output$plot2 <- renderPlotly({
    p <- plot_ly() %>% add_trace(x = space$x_coord, y = space$y_coord, z = space$z_coord, 
                                 mode='markers',
                                 hoverinfo = 'text',
                                 marker=list(
                                   color = space$cluster_default_rate, 
                                   colorscale='RdYlBu',
                                   colorbar=list(title='Default rate')
                                 ),
                                 text = ~paste('</br>Cluster', space$cluster, 
                                               '</br>Cluster default rate:', space$cluster_default_rate, 
                                               '</br>Purpose', space$purpose), # TODO display all relevant data,
                                 type = "scatter3d"
    ) 
    
    new_data_point <- space[space$test == T, ][5,]
    
    p %>% layout(scene = list(
      aspectratio = list(
        x = 1,
        y = 1,
        z = 1
      ),
      camera = list(
        center = list(
          x = 0,
          y = 0,
          z = 0
        ),
        eye = list(
          x = 1.96903462608,
          y = -1.09022831971,
          z = 0.405345349304
        ),
        up = list(
          x = 0,
          y = 0,
          z = 1
        )
      ), annotations = list(list(
        showarrow = T,
        # datapoint coordinates:
        x = new_data_point$x_coord,
        y = new_data_point$y_coord,
        z = new_data_point$z_coord,
        text = "New loan application",
        xshift = 0,
        opacity = 1
      ))))
  })
  
  output$brush_info <- renderDataTable({
    brushedPoints(tsne_data, input$plot1_brush)
  })
}

shinyApp(ui, server)
```
